Large commit with 6 files from MTalk a prior project of mine.
* .gitignore: Added both vim and emacs temp files

* network_utils.cpp:
(get_host_by_name): return the network address of a host as a string
* network_utils.h: define the above .cpp
* socket.cpp:
(initialize_winsock):
(Socket): This is a class, look for the definition in the header file.
(bind): bind to a local port.
(create_address): get us an address we can use, just a pretty wrapper
over some of the low level C stuff.
(connect): connect to a remote host.
(listen): look for connections
(close): close the socket handle, set our socket_id to -1 to indicate this.
(send): send all the data we have to send, keep sending until
everything is sent through the socket
(accept): accept a new connection
(set_nonblocking): set our socket to not block
(get_socketid): Get filehandle of socket.
(recv): return the *non buffered* data from the remote connection. We
cannot buffer without an idea of the protocol that we are handling so
this is left up to class inheriters or users
* socket.h: header file.
* string_utils.cpp:
(Utils::tokenize): tokenize a string, handwritten, probably something
better out there.
* string_utils.h: our function definition.
* defines.h: Various macro definitions such as program version. None
actually in